Add loss function and optimizer. Change data type of input IQ vectors from float64 to float32 just because.

+++ src/chirpgen/ChirpClassifierModel.py
@@ -14,8 +14,8 @@ class ChirpClassifierModel(nn.Module):
     # self.num_hidden_feats = self.num_in_feat + int(self.num_in_feat * 0.5)
     self.num_hidden_feats = 1024
     print(f"num_hidden_feats: {self.num_hidden_feats}")
-    self.layer1 = nn.Linear(in_features=self.num_in_feats, out_features=self.num_hidden_feats, dtype=torch.float64)
-    self.layer2 = nn.Linear(in_features=self.num_hidden_feats, out_features=1, dtype=torch.float64)
+    self.layer1 = nn.Linear(in_features=self.num_in_feats, out_features=self.num_hidden_feats, dtype=torch.float32)
+    self.layer2 = nn.Linear(in_features=self.num_hidden_feats, out_features=1, dtype=torch.float32)
 
   def forward(self, batch_iq_tensors):
     batch_sz = len(batch_iq_tensors)
